+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * Add validation for the `claimingInterval` field. Refs UIOR-1192.
 * *BREAKING* Add settings components for custom fields. Refs UIOR-1177.
 * Bump up okapi interfaces for `pieces` (2.0 3.0). Refs UIOR-1212.
+* Opening and editing POs with location-restricted funds (FE). Refs UIOR-1184.
 
 ## [5.0.1](https://github.com/folio-org/ui-orders/tree/v5.0.1) (2023-11-08)
 [Full Changelog](https://github.com/folio-org/ui-orders/compare/v5.0.0...v5.0.1)